requirements.txt
# Core ML libraries
numpy==1.26.0
pandas==2.0.3
scikit-learn==1.3.0

# Data validation
great-expectations==0.18.8

# Experiment tracking
mlflow==2.10.2

# Data versioning
dvc==3.48.4

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# ONNX inference (ADDED — безопасно)
onnx==1.20.0
onnxruntime==1.23.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Code quality
black==23.7.0
flake8==7.1.1
isort==5.12.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
joblib==1.3.2
requests==2.31.0
torch
